removes sklearn pipeline [Python]
COMMENTS:
It is more instructive to look at the individual steps from an
academic point of view. The pipeline is more suitable for data
scientists working in a production environment.

Plots the scores for the train and test sets to illustrate that
although the training set is described very well for degrees
higher than or equal to two, the test scores are poor.