adds random forest regressor example [Python]
COMMENTS:
The advantage of using the random forests is that it helps identify
the unimportant features of the dataset.

After extraction of the unimportant features and standarization, the
best fit was obtained for a simple linear regression with polynomial
features of a second degree polynomial.

The work is based on reference [3] (see the synopsis).