updates ANN and splits into train test [CXX]
COMMENTS:
The ANN construction now closely follows reference [2], where the
bias is a scalar quantity not a vector (you are encouraged to
look at the source code of the author).

The bias vector lead to overfitting. The ANN would closely
describe the train set but fail to do so for the test set.
The issue is that the size of the bias vector is determined by the
number of measurements. Since the train and test sets differ in
size it is clear that the optimized bias vector is incompatible
for the test set from a linear algebra viewpoint.

Now the ANN is just able to capture the qualitative behavior of
the sets; however, this is to be expected for such a simple ANN that
consists of just the input layer. It would be interesting to see how
a robust ANN would perform.